Initialize incoming items buffer with batchlength capacity
We can safely assume the full buffer is needed, so we can as well
allocate enough capacity. This makes append significantly faster as
slice already has enough capacity.

This will cause a small amount of unnecessary overhead if we don't
fill up the buffer before maximum batch wait time is exceeded.
However, in that case when the buffer is copied before sending it
over the channel, only len, not cap is allocated.