Add a simple receiver server
This can be used to receive and write out messages from the sender.
The general idea is to run the sender behind unreliable internet
connection and receiver on a reliable server. When the internet
connection is up, sender's TCP mode can be used to transmit all
messages to the receiver, which will then forward those to a file
or to stdout (or to any other destination, depending on the
configuration).

TLS is to-be-added feature; any encrypted tunnel - ipsec, ssh
tunnel etc. can be used to provide encryption and authentication.